Step 5: One-Way Data Binding
Here, data is transported in one direction only: from the model, through the binding instance to the consumer (usually the property of a control), but never in the other direction. In this example, we change the previous example to use one-way data binding. This illustrates how the flow of data from the user interface back to the model can be switched off if required.