Step 14: Expression Binding
Expression binding allows you to display a value on the screen that has been calculated from values found in some model object. This way simple formatting or calculations can be inserted directly into the data binding string. In this example, we change the color of the price depending on whether it is above or below some arbitrary threshold. The threshold value is also stored in the JSON model